refactor: provider폴더 위치 변경
