+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { RouterProvider } from "react-router-dom";
 import { router } from "./router/router";
-import { AppProvider } from "./pages/providers     /AppProvider";
+import { AppProvider } from "./providers/AppProvider";
 
 const App = () => {
   return (
